shopifyapp/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.36
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
sendgrid==6.9.7
celery==5.2.6
redis==4.3.1
pytest==6.2.5
pipenv==2022.5.2
itsdangerous==2.1.2
boto3==1.23.1
PyJWT==2.3.0
